requirements.txt
# Core dependencies
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Vector store
chromadb>=0.4.0

# Document processing
pypdf2>=3.0.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0

